too much stuff to describe...

+++ source_separation/recurrent_reconstrcution.py
@@ -29,7 +29,7 @@ def create_batches(sample, mask, win_len):
 
     return train_x, mask_x, mask_y
 
-def pad_sequences(sequences, max_len, dtype='int32', padding='post', truncating='post', transpose=True, value=0.):
+def pad_sequences(sequences, max_len, dtype='float32', padding='post', truncating='post', transpose=True, value=0.):
     # (nb_samples, max_sample_length (samples shorter than this are padded with zeros at the end), input_dim)
     nb_samples = len(sequences)
     x = (np.ones((nb_samples, max_len, sequences[0].shape[1])) * value).astype(dtype)
